Refactor rabbitmq modules so verifier & tests pass
