+++ instrumentation/rabbit-amqp-1.7.2/build.gradle
@@ -1,0 +1,19 @@
+dependencies {
+    implementation(project(":agent-bridge"))
+    implementation("com.rabbitmq:amqp-client:1.7.2")
+    testImplementation("io.arivera.oss:embedded-rabbitmq:1.4.0")
+    testImplementation("org.slf4j:slf4j-simple:1.7.30")
+}
+
+jar {
+    manifest { attributes 'Implementation-Title': 'com.newrelic.instrumentation.rabbit-amqp-1.7.2' }
+}
+
+verifyInstrumentation {
+    passesOnly 'com.rabbitmq:amqp-client:[1.7.2,2.4.1)'
+}
+
+site {
+    title 'RabbitAMQP'
+    type 'Messaging'
+}
