+++ instrumentation/rabbit-amqp-1.7.2/src/integration-test/java/RabbitMQTest_Integration.java
