+++ instrumentation/rabbit-amqp-1.7.2/src/main/java/com/nr/agent/instrumentation/rabbitamqp27/RabbitAMQPMetricUtil.java
