+++ instrumentation/rabbit-amqp-1.7.2/src/main/java/com/rabbitmq/client/FileProperties_Instrumentation.java
