+++ instrumentation/rabbit-amqp-1.7.2/src/main/java/com/rabbitmq/client/QueueingConsumer_Instrumentation.java
