+++ instrumentation/rabbit-amqp-1.7.2/src/test/java/com/nr/agent/instrumentation/rabbitamqp27/RabbitMQTest.java
