+++ instrumentation/rabbit-amqp-2.4.1/build.gradle
@@ -1,0 +1,19 @@
+dependencies {
+    implementation(project(":agent-bridge"))
+    implementation("com.rabbitmq:amqp-client:2.4.1")
+    testImplementation("io.arivera.oss:embedded-rabbitmq:1.4.0")
+    testImplementation("org.slf4j:slf4j-simple:1.7.30")
+}
+
+jar {
+    manifest { attributes 'Implementation-Title': 'com.newrelic.instrumentation.rabbit-amqp-2.4.1' }
+}
+
+verifyInstrumentation {
+    passesOnly 'com.rabbitmq:amqp-client:[2.4.1,2.5.0)'
+}
+
+site {
+    title 'RabbitAMQP'
+    type 'Messaging'
+}
