+++ instrumentation/rabbit-amqp-2.7.0/build.gradle
@@ -1,0 +1,20 @@
+dependencies {
+    implementation(project(":agent-bridge"))
+    implementation("com.rabbitmq:amqp-client:3.5.0")
+
+    testImplementation("io.arivera.oss:embedded-rabbitmq:1.4.0")
+    testImplementation("org.slf4j:slf4j-simple:1.7.30")
+}
+
+jar {
+    manifest { attributes 'Implementation-Title': 'com.newrelic.instrumentation.rabbit-amqp-2.7.0' }
+}
+
+verifyInstrumentation {
+    passesOnly 'com.rabbitmq:amqp-client:[2.7.0,5.0.0.RC1)'
+}
+
+site {
+    title 'RabbitAMQP'
+    type 'Messaging'
+}
