+++ settings.gradle
@@ -296,8 +296,10 @@ include 'instrumentation:r2dbc-mysql'
 include 'instrumentation:r2dbc-postgresql-0.9.0'
 include 'instrumentation:r2dbc-postgresql-0.9.2'
 include 'instrumentation:r2dbc-mssql'
-include 'instrumentation:rabbit-amqp-2.7'
-include 'instrumentation:rabbit-amqp-3.5.0'
+include 'instrumentation:rabbit-amqp-1.7.2'
+include 'instrumentation:rabbit-amqp-2.4.1'
+include 'instrumentation:rabbit-amqp-2.5.0'
+include 'instrumentation:rabbit-amqp-2.7.0'
 include 'instrumentation:rabbit-amqp-5.0.0'
 include 'instrumentation:reactor-3.3.0'
 include 'instrumentation:resin-3'
